+++ kmc_system.h
@@ -41,7 +41,7 @@ class class_system{
 		// Functions
 		void init_states_array(int nVset, double compA, int* const states000);
 		void read_restart(char name_restart[], int* const states000, int &ts_initial, double &time_initial);
-		void write_conf(int timestep, double time, int *ptr_states);
+		void write_conf(long long int timestep, double time, int *ptr_states);
 
 	private:
 		int       nx, ny, nz;
